layout refactor, now navbar controls scrollTopButton to improve the page performance, all images compresed to save mobile bandwidth and make the web load faster, little modification on paddingbottom in MyDescription
