+++ components/TechStack/TechStack.module.scss
@@ -5,7 +5,7 @@
   padding-top: 4rem;
 
   @include BreakPoint.mobile {
-    padding-bottom: 110px;
+    padding-bottom: 120px;
   }
 
   > ul {
